# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for UniSphere university social platform",
        "title": "UniSphere API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://www.unisphere.com/support",
            "email": "support@unisphere.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/department-instructors/{departmentId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of instructors for a specific department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instructors"
                ],
                "summary": "Get instructors by department",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int64",
                        "description": "Department ID",
                        "name": "departmentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instructors retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.InstructorResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid department ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Department not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/instructors/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed profile information for the currently authenticated instructor (Requires instructor role)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instructors"
                ],
                "summary": "Get instructor profile",
                "responses": {
                    "200": {
                        "description": "Instructor profile retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.InstructorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not an instructor or token invalid",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Instructor not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/instructors/title": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the academic title for the currently authenticated instructor (Requires instructor role)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instructors"
                ],
                "summary": "Update instructor title",
                "parameters": [
                    {
                        "description": "New title information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.UpdateTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instructor title updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format or validation error (e.g., empty title)",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not an instructor or token invalid",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Instructor not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/instructors/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a specific instructor by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instructors"
                ],
                "summary": "Get instructor by ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int64",
                        "description": "Instructor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instructor retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.InstructorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid instructor ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Instructor not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_yigit_unisphere_internal_app_models_dto.APIResponse": {
            "description": "Generic response structure for all API endpoints",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload - specific to each endpoint"
                },
                "timestamp": {
                    "description": "Timestamp of when the response was generated",
                    "type": "string",
                    "example": "2025-04-23T12:01:05.123Z"
                }
            }
        },
...
        "github_com_yigit_unisphere_internal_app_models_dto.ErrorResponse": {
            "description": "Generic response structure for all API endpoints",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload - specific to each endpoint"
                },
                "timestamp": {
                    "description": "Timestamp of when the response was generated",
                    "type": "string",
                    "example": "2025-04-23T12:01:05.123Z"
                }
            }
        },
...
        "github_com_yigit_unisphere_internal_app_models_dto.InstructorResponse": {
            "type": "object",
            "properties": {
                "departmentId": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "profilePhotoUrl": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "github_com_yigit_unisphere_internal_app_models_dto.UpdateTitleRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Professor"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token (without Bearer prefix)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}